c++ project
Departmental Store Tracker Application in C++ The Project will use the concept of file handling and structures and all other concepts of c++ to store the details of departmental store across the country. The user can search the departmental store by entering the state and city. The admin can insert and delete the new and existing stores respectively.